remove critical errors for normal errors into applyRequest method
