+++ x/launch/types/expected_keepers.go
@@ -5,6 +5,7 @@ import (
 )
 
 type ProfileKeeper interface {
+	HasCoordinator(ctx sdk.Context, id uint64) bool
 	CoordinatorIDFromAddress(ctx sdk.Context, address string) (id uint64, found bool)
 	GetCoordinatorAddressFromID(ctx sdk.Context, id uint64) (address string)
 }
